Don't use unicode_literals in tests.
Some of the tests use strings which should be native strings: str on
python2 and python3, and it's not possible to (easily) have native
string literals with unicode_literals on.

Additionally this means I can get rid of the stupid compatibility hack I
added to _libpassacre to get tests passing on python 2.