Fix building all the extension crap.
Holy crap this is getting awful. The build was failing before because
the build_ext check was getting run before build_ext.finalize_options,
so it decided that it shouldn't build an extension. So, put it back in
the build command, which delegates to build_ext, which delegates to
build_libpassacre, ...