Allow persistence of the scrypted password.
I'm going to have to handle this /very/ carefully, but the intent here
is to be able to persist the scrypt output somewhere so that it doesn't
have to be calculated every time. This should only ever be done on
implicitly trusted hardware, so probably not something like a laptop.
It's more intended for passacre-dedicated hardware.

The API might have to change, but we'll see.