Fix the travis builds.
This will /actually/ use clang for coverage analysis. It turns out that
travis's builtin clang is turbofucked and not actually usable, and you
have to install the clang-3.4 package to get something usable. After
that, it's okay to use clang as the linker, and using clang as the
linker with --coverage produces coverage-tracking shared objects!

--coverage is used instead of the -f flags because it enables everything
you need easier. Also it turns out that distutils will use CFLAGS for
linking too so LDFLAGS isn't necessary.