Do aligned allocations for keccak.
It might actually matter, even if it kind of didn't look like it.

In trying to use aligned_alloc/posix_memalign, I discovered that nothing
was actually using the HAVE_POSIX_MEMALIGN that I'd added to CMake, so I
fixed that.

Originally I thought that using -std=c11 with clang would allow use of
aligned_alloc, but that appears to not be the case. Either way, CMake
now determines whether it should use C11 or C99 instead of assuming C99.