Add real error handling to libpassacre.
The biggest change is the addition of a passacre_error function to
libpassacre, which will explain what the various errors mean. It writes
to a string buffer with a stringification of the error code.

An important piece of this is trapping panics from bubbling up out of
rust into C. In that case, the text of the panics is recorded to be
pulled out with passacre_error later.

The other important piece is adding a real error enum to passacre.rs
instead of just using () for 'something went wrong'. Errors are now more
explicitly specified, which required changing types around.

The downside to all of this is that it requires some unstable rust APIs
at the moment. So, users will need beta channel rust instead of stable
channel rust. Not a big deal, as these things should settle down at some
point, but it is slightly unfortunate.

Some todos:

 - Trap panics in passacre_error too.
 - Add tests around passacre_error, which might involve trying to figure
   out how to trigger panics from the C API.