No more C API. Good riddance.
It was unsafe and crappy. Exposing capnproto RPC over a socket is 900
fewer lines of code and actually safe. Also fix some other unsafe bits,
like using mem::uninitialized.

Tests are not all passing yet, but the bulk of them are. The last few
are just related to remote error handling (which isn't done) and
scrypt (which was never stable).